simulador_backend/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-rest==0.8.7
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
